Bugfix: "missing" textures (part 1)
An extra, empty slot made appear some objects as pink, but no texture was actually missing. Most of those being on LOD it is easy to miss if not seen from the correct distance. No error on prefab, it happened to the map itself.
- rockBig2_lod0
- zenkei_stonePath (various)

Shader was error:
- now restored to "UniversalRenderPipeline/Lit" on the following materials
--- MarketSignPaper3 (fireworks)
--- wheat_cutout
--- Foliage_Atlas
--- Starfish
- now restored to "UniversalRenderPipeline/Particles/Lit" on the following materials
--- waterSplash3
- Albedo texture not found ( /!\ need fix later /!\ )
--- reed

Other observations, need import of "Landscape" before further investigation:
- Sand albedo/normal missing from terrain brush (the checkerboard area)
- Lot of missing "brushes" in terrain (trees and details)